refactor: reorganize imports and add NumberInput component for better input handling
