+++ frontend/components/ui/NumberInput.tsx
